Implement KWargs for ActionView's Select Helper
I made a quick and dirty workaround for
test_select_with_multiple_and_disabled_to_add_disabled_hidden_input in
order to make it pass. Please review the changes and I am sure this
workorund can be re-written in a better way. I marked it with FIXME in
the form_options_helper file.